fix: default to pages router to prep for incremnetal builds
